Move app context in its own module
